Instructions
ADD, ADC

+++ registers.h
@@ -3,5 +3,5 @@
 uint8_t R[32];
 
 struct SREG{
-    int I,T,H,S,V,N,Z,C;
+    uint8_t I,T,H,S,V,N,Z,C;
 }SREG;
